[YS-416] refactor: 탈퇴 시 유의사항 컴포넌트 분리
